organize & add stack to exercisesTwo folder
